feat: update interfaces for LocatorType & ElementAttributes
